requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
